requirements.txt
Click==7.0
Cython==0.29.4
Flask==1.0.2
Flask-API==1.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.1
opencv-python==4.0.0.21
Pillow==5.4.1
pytesseract==0.2.6
tesseract==0.1.3
Werkzeug==0.14.1